refactor: add and use OAuthService interface
